+++ writing/nvim/lua/plugins/lsp.lua
@@ -2,9 +2,9 @@ return {
 	{
 		"neovim/nvim-lspconfig",
 		ft = { "lua", "python", "sh", "bash" },
-		dependencies = { "hrsh7th/cmp-nvim-lsp" },
+		dependencies = { "saghen/blink.cmp" },
 		config = function()
-			local capabilities = require("cmp_nvim_lsp").default_capabilities()
+			local capabilities = require("blink.cmp").get_lsp_capabilities()
 
 			vim.lsp.config("lua_ls", { capabilities = capabilities })
 			vim.lsp.config("pyright", { capabilities = capabilities })